requirements.txt
streamlit
requests
streamlit-autorefresh
aeroapi-python